backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis / task queue
redis==5.0.4
celery==5.4.0
celery[redis]==5.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP
httpx==0.27.0

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# LLM (compression agent)
openai==1.30.1
tiktoken==0.7.0

# Notifications
slack-sdk==3.27.2

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
